fix: auth and onboarding
