--- src/features/home/model/effects/useCenterRedirect.ts
@@ -1,9 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-
-export function useCenterRedirect(checkAndRedirect: () => Promise<void>): void {
-  useEffect(() => {
-    void checkAndRedirect()
-  }, [checkAndRedirect])
-}
