--- src/features/questions-upload/model/effects/useInitialLoad.ts
@@ -1,15 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-
-export function useInitialLoad(
-  subjectId: string,
-  sessionId: string,
-  loadData: (bypassCache?: boolean) => Promise<void>
-): void {
-  useEffect(() => {
-    if (subjectId && sessionId) {
-      loadData()
-    }
-  }, [subjectId, sessionId, loadData])
-}
