--- src/features/session-questions-upload/model/effects/useVisibilityRefresh.ts
@@ -1,19 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-
-export function useVisibilityRefresh(
-  sessionId: string,
-  loadSessionAndStats: (bypassCache?: boolean) => Promise<void>
-): void {
-  useEffect(() => {
-    const handleVisibilityChange = () => {
-      if (document.visibilityState === 'visible' && sessionId) {
-        loadSessionAndStats(true)
-      }
-    }
-
-    document.addEventListener('visibilitychange', handleVisibilityChange)
-    return () => document.removeEventListener('visibilitychange', handleVisibilityChange)
-  }, [sessionId, loadSessionAndStats])
-}
