+++ src/features/settings/model/scoreConfigurations/useScoreConfigurations.ts
@@ -1,14 +1,26 @@
 'use client'
 
+import { useCallback, useEffect } from 'react'
 import type { Center } from '@/entities/center'
+import { createLoadScoreConfigurations } from './handlers/createLoadScoreConfigurations'
 import type { ScoreConfigurationsController } from './types'
-import { useScoreConfigurationLoad } from './effects/useScoreConfigurationLoad'
-import { useScoreConfigurationState } from './effects/useScoreConfigurationState'
+import { useScoreConfigurationState } from './useScoreConfigurationState'
 import { useScoreConfigurationsController } from './useScoreConfigurationsController'
 
 export function useScoreConfigurations(centers: Center[]): ScoreConfigurationsController {
   const state = useScoreConfigurationState()
-  const reload = useScoreConfigurationLoad(state)
+
+  const { setScoreConfigurations, setIsLoadingScores, setScoreError } = state
+
+  const reload = useCallback(
+    () =>
+      createLoadScoreConfigurations(setScoreConfigurations, setIsLoadingScores, setScoreError)(),
+    [setScoreConfigurations, setIsLoadingScores, setScoreError]
+  )
+
+  useEffect(() => {
+    void reload()
+  }, [reload])
 
   return {
     ...state,
