+++ src/features/settings/model/scoreConfigurations/useScoreConfigurationsFormActions.ts
@@ -4,10 +4,10 @@ import { useCallback } from 'react'
 import type { FormEvent } from 'react'
 import type { Center } from '@/entities/center'
 import type { ScoreConfiguration } from '@/entities/score-configuration'
-import { createHandleCancelEdit } from '../handlers/createHandleCancelEdit'
-import { createHandleCreate } from '../handlers/createHandleCreate'
-import { createHandleEdit } from '../handlers/createHandleEdit'
-import { createHandleUpdate } from '../handlers/createHandleUpdate'
+import { createHandleCancelEdit } from './handlers/createHandleCancelEdit'
+import { createHandleCreate } from './handlers/createHandleCreate'
+import { createHandleEdit } from './handlers/createHandleEdit'
+import { createHandleUpdate } from './handlers/createHandleUpdate'
 import type { ScoreConfigurationState } from './useScoreConfigurationState'
 
 export function useScoreConfigurationsFormActions(
