--- src/processes/center-check/index.ts
@@ -1,2 +1,0 @@
-export { checkCenterExists } from './lib/checkCenter'
-export { cache } from './lib/cache'
