--- src/processes/center-check/lib/cache.ts
@@ -1,37 +1,0 @@
-interface CacheEntry<T> {
-  value: T
-  expiresAt: number
-}
-
-class SimpleCache<T> {
-  private cache = new Map<string, CacheEntry<T>>()
-
-  set(key: string, value: T, ttlMs: number): void {
-    this.cache.set(key, {
-      value,
-      expiresAt: Date.now() + ttlMs,
-    })
-  }
-
-  get(key: string): T | null {
-    const entry = this.cache.get(key)
-    if (!entry) return null
-
-    if (Date.now() > entry.expiresAt) {
-      this.cache.delete(key)
-      return null
-    }
-
-    return entry.value
-  }
-
-  clear(key?: string): void {
-    if (key) {
-      this.cache.delete(key)
-    } else {
-      this.cache.clear()
-    }
-  }
-}
-
-export const cache = new SimpleCache<boolean>()
